+++ jpa/plugins/org.eclipse.jpt.core/src/org/eclipse/jpt/core/internal/context/orm/GenericOrmColumn.java
@@ -38,11 +38,11 @@ public class GenericOrmColumn extends AbstractOrmBaseColumn<XmlColumn> implement
 		setSpecifiedScale(oldColumn.getSpecifiedScale());
 	}
 
-	public Integer getLength() {
-		return (this.getSpecifiedLength() == null) ? getDefaultLength() : this.getSpecifiedLength();
+	public int getLength() {
+		return (this.getSpecifiedLength() == null) ? getDefaultLength() : this.getSpecifiedLength().intValue();
 	}
 
-	public Integer getDefaultLength() {
+	public int getDefaultLength() {
 		return Column.DEFAULT_LENGTH;
 	}
 
@@ -74,11 +74,11 @@ public class GenericOrmColumn extends AbstractOrmBaseColumn<XmlColumn> implement
 		firePropertyChanged(SPECIFIED_LENGTH_PROPERTY, oldSpecifiedLength, newSpecifiedLength);
 	}
 
-	public Integer getPrecision() {
-		return (this.getSpecifiedPrecision() == null) ? getDefaultPrecision() : this.getSpecifiedPrecision();
+	public int getPrecision() {
+		return (this.getSpecifiedPrecision() == null) ? getDefaultPrecision() : this.getSpecifiedPrecision().intValue();
 	}
 
-	public Integer getDefaultPrecision() {
+	public int getDefaultPrecision() {
 		return Column.DEFAULT_PRECISION;
 	}
 	
@@ -110,11 +110,11 @@ public class GenericOrmColumn extends AbstractOrmBaseColumn<XmlColumn> implement
 		firePropertyChanged(SPECIFIED_PRECISION_PROPERTY, oldSpecifiedPrecision, newSpecifiedPrecision);
 	}
 
-	public Integer getScale() {
-		return (this.getSpecifiedScale() == null) ? getDefaultScale() : this.getSpecifiedScale();
+	public int getScale() {
+		return (this.getSpecifiedScale() == null) ? getDefaultScale() : this.getSpecifiedScale().intValue();
 	}
 
-	public Integer getDefaultScale() {
+	public int getDefaultScale() {
 		return Column.DEFAULT_SCALE;
 	}
